Add functionality for x86 ubuntu system
