+++ calc_arrays.h
@@ -1,9 +1,12 @@
+#pragma once
+
 #include <Eigen/Dense>
 #include <iostream>
 
 using Eigen::MatrixXd;
 
-struct CameraInfo {
+struct CameraInfo
+{
     double yaw;
     double pitch;
     double roll;
@@ -13,7 +16,8 @@ struct CameraInfo {
     double opticalCenterY;
     double cameraHeight;
 };
-struct IpmInfo{
+struct IpmInfo
+{
     double inputWidth;
     double inputHeight;
     double left;
@@ -21,7 +25,7 @@ struct IpmInfo{
     double top;
     double bottom;
 };
-MatrixXd GetVanishingPoint(const CameraInfo& cameraInfo);
+MatrixXd GetVanishingPoint(const CameraInfo &cameraInfo);
 MatrixXd TransformGround2Image(MatrixXd xyGrid, CameraInfo cameraInfo);
-MatrixXd TransformImage2Ground(const MatrixXd& uvLimits, const CameraInfo& cameraInfo);
-MatrixXd GetMappingArrays(const CameraInfo& cameraInfo, IpmInfo& ipmInfo);
+MatrixXd TransformImage2Ground(const MatrixXd &uvLimits, const CameraInfo &cameraInfo);
+MatrixXd GetMappingArrays(const CameraInfo &cameraInfo, IpmInfo &ipmInfo);
